+++ .github/workflows/graalvm.yml
@@ -44,7 +44,7 @@ jobs:
           GRADLE_ENTERPRISE_CACHE_USERNAME: ${{ secrets.GRADLE_ENTERPRISE_CACHE_USERNAME }}
           GRADLE_ENTERPRISE_CACHE_PASSWORD: ${{ secrets.GRADLE_ENTERPRISE_CACHE_PASSWORD }}
           GIT_COMMIT_HASH: ${{ github.sha }}
-          OCI_REGION: ${{ secrets.OCI_REGION }}
+          OCI_CLI_REGION: ${{ secrets.OCI_REGION }}
           JOB_ID: ${{ format('{0}-graalvm-java-{1}', github.run_id, matrix.java) }}
           OCI_TENANCY_NAME: ${{ secrets.OCI_TENANCY_NAME }}
           JAVA_VERSION: ${{ matrix.java }}
