+++ .github/workflows/gradle.yml
@@ -42,7 +42,7 @@ jobs:
           GRADLE_ENTERPRISE_ACCESS_KEY: ${{ secrets.GRADLE_ENTERPRISE_ACCESS_KEY }}
           GRADLE_ENTERPRISE_CACHE_USERNAME: ${{ secrets.GRADLE_ENTERPRISE_CACHE_USERNAME }}
           GRADLE_ENTERPRISE_CACHE_PASSWORD: ${{ secrets.GRADLE_ENTERPRISE_CACHE_PASSWORD }}
-          OCI_REGION: ${{ secrets.OCI_REGION }}
+          OCI_CLI_REGION: ${{ secrets.OCI_REGION }}
           OCI_TENANCY_NAME: ${{ secrets.OCI_TENANCY_NAME }}
           AUTH_TOKEN: ${{ secrets.OCI_AUTH_TOKEN }}
           OCIR_USERNAME: ${{ secrets.OCIR_USERNAME }}
